requirements.txt
flask==2.3.3
python-dotenv==1.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.13
supabase==2.13.0
cloudinary==1.36.0
Pillow==10.0.0
uuid==1.30
ultralytics>=8.0.0  # For SAM
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0  # For Grounding DINO
matplotlib>=3.7.0
numpy>=1.24.0
requests>=2.31.0
pytesseract>=0.3.10  # For OCR